Added fields to the buttons to keep track of the amount a button was pressed on fragment_main.xml

+++ app/src/main/java/com/example/jelliott/idealapplication/Family.java
@@ -85,7 +85,7 @@ public class Family {
             this.familyWealth = familyWealthA;
         }
 
-        public void setFamilyInfluence(int familyInfluenceA) {
+        public void setFamilyInfluence(double familyInfluenceA) {
             this.familyInfluence = familyInfluenceA;
         }
